+++ jdbc/src/main/java/nextstep/jdbc/ExecuteStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc.resultset;
+package nextstep.jdbc;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
