+++ jdbc/src/main/java/nextstep/jdbc/PreparedStatementCreator.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc.resultset;
+package nextstep.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
